Added API Util input validation
