+++ server/errors/APIError.js
@@ -1,0 +1,8 @@
+class APIError extends Error {
+    constructor(message) {
+        super(message)
+        this.name = 'APIError'
+    }
+}
+
+module.exports = { APIError }
