--- server/utils/ebay/EbayAPIUtils.js
@@ -1,59 +1,0 @@
-const { json } = require('body-parser')
-const EbayAuthToken = require('ebay-oauth-nodejs-client')
-require('dotenv').config({
-    path: `${__dirname}/../../.env`
-})
-
-const EBAY_API_BROWSE_URL = process.env.EBAY_API_BROWSE_URL
-const generateClientAccessToken = async () => {
-    try {
-        const ebayAuthToken = new EbayAuthToken({
-            clientId: process.env.EBAY_CLIENT_ID,
-            clientSecret: process.env.EBAY_CLIENT_SECRET,
-            redirectUri: process.env.EBAY_REDIRECT_URL,
-        })
-        const environment = 'PRODUCTION'
-        const scopes = ['https://api.ebay.com/oauth/api_scope']
-        const response = await ebayAuthToken.getApplicationToken(environment, scopes)
-        const data = await JSON.parse(response)
-        return data['access_token']
-    } catch (error) {
-        console.error(error)
-        throw new Error(error)
-    }
-}
-
-const getListings = async (keyword, limit, category) => {
-    const clientAccessToken = await generateClientAccessToken()
-    const response = await fetch(`${EBAY_API_BROWSE_URL}/item_summary/search?q=${keyword}&limit=${limit}&category_ids=${category}&fieldgroups=EXTENDED`, {
-        method: 'GET',
-        headers: {
-            'Authorization': `Bearer ${clientAccessToken}`
-        }
-    })
-    return await response.json()
-}
-
-const getItem = async (itemHref) => {
-    const clientAccessToken = await generateClientAccessToken()
-    const response = await fetch(itemHref, {
-        method: 'GET',
-        headers: {
-            'Authorization': `Bearer ${clientAccessToken}`
-        }
-    })
-    return await response.json()
-}
-
-const checkRateLimit = async (api) => {
-    const clientAccessToken = await generateClientAccessToken()
-    const response = await fetch(`https://api.ebay.com/developer/analytics/v1_beta/rate_limit/?api_name=${api}`, {
-        method: 'GET',
-        headers: {
-            'Authorization': `Bearer ${clientAccessToken}`
-        }
-    })
-    return await response.json()
-}
-
-module.exports = { getListings, getItem, checkRateLimit }
